requirements.txt
Django==5.1.7
dj-rest-auth==7.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-allauth==65.5.0
django-cors-headers==4.7.0
Pillow==11.1.0
python-decouple==3.8
requests==2.32.3
social-auth-app-django==5.4.3
